refactor: Update webview start and timeline views
